started fill/growingEnd cleanup

+++ GW2EIEvtcParser/EIData/CombatReplay/Decorations/IconOverheadDecoration.cs
@@ -6,7 +6,7 @@ namespace GW2EIEvtcParser.EIData
     internal class IconOverheadDecoration : IconDecoration
     {
 
-        public IconOverheadDecoration(string icon, int pixelSize, float opacity, (int start, int end) lifespan, AgentConnector connector) : base(icon, pixelSize, Math.Min((int)connector.Agent.HitboxWidth, 250), opacity, lifespan, connector)
+        public IconOverheadDecoration(string icon, int pixelSize, float opacity, (long start, long end) lifespan, AgentConnector connector) : base(icon, pixelSize, Math.Min((int)connector.Agent.HitboxWidth, 250), opacity, lifespan, connector)
         {
         }
